moving torch hub methods to torch utils
